pyproject.toml
[tool.poetry]
name = "application-for-analyzing-banking-transactions"
version = "0.1.0"
description = ""
authors = ["Илья Гирько <Gggggirr@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
openpyxl = "^3.1.5"
requests = "^2.32.3"
pandas = "^2.2.2"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"
isort = "^5.13.2"
mypy = "^1.11.2"
black = "^24.8.0"


[tool.poetry.group.pytest.dependencies]
pytest-cov = "^5.0.0"
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
disallow_untyped_defs = true
warn_return_any = true
exclude = "venv"


[tool.black]
line-length = 119
exclude = """(/(|/.git)/)"""


[tool.isort]
line_length = 119